Move cross_variation module into model_selection
