+++ sklearn/linear_model/coordinate_descent.py
@@ -16,7 +16,7 @@ from .base import LinearModel, _pre_fit
 from ..base import RegressorMixin
 from .base import center_data, sparse_center_data
 from ..utils import array2d, atleast2d_or_csc
-from ..cross_validation import _check_cv as check_cv
+from ..model_selection.partition import _check_cv as check_cv
 from ..externals.joblib import Parallel, delayed
 from ..externals import six
 from ..externals.six.moves import xrange
