+++ sklearn/model_selection/__init__.py
@@ -1,0 +1,17 @@
+from .partition import LeaveOneOut, LeavePOut, KFold, \
+        StratifiedKFold, LeaveOneLabelOut, LeavePLabelOut, \
+        ShuffleSplit, StratifiedShuffleSplit, train_test_split, check_cv
+from .validate import cross_val_score, permutation_test_score
+
+__all__ = ['KFold',
+           'LeaveOneLabelOut',
+           'LeaveOneOut',
+           'LeavePLabelOut',
+           'LeavePOut',
+           'ShuffleSplit',
+           'StratifiedKFold',
+           'StratifiedShuffleSplit',
+           'check_cv',
+           'cross_val_score',
+           'permutation_test_score',
+           'train_test_split']
